chore: breakout config backup
