+++ internal/agents/inventory.go
@@ -16,11 +16,8 @@ import (
 	"net/url"
 	"os"
 	"os/exec"
-	"path/filepath"
 	"strings"
-	"time"
 
-	"github.com/circonus/agent-manager/internal/config/defaults"
 	"github.com/circonus/agent-manager/internal/config/keys"
 	"github.com/circonus/agent-manager/internal/registration"
 	"github.com/rs/zerolog/log"
@@ -264,61 +261,4 @@ func registerAgents(ctx context.Context, c InstalledAgents) error {
 	log.Debug().Str("resp", string(body)).Msg("response")
 
 	return nil
-}
-
-func backupConfigs(name string, configs map[string]string) {
-	ts := time.Now().Format("20060102_150405")
-	baseDir := filepath.Join(defaults.EtcPath, "configs", name)
-
-	if err := os.MkdirAll(baseDir, 0700); err != nil {
-		if !errors.Is(err, os.ErrExist) {
-			log.Error().Err(err).Str("path", baseDir).Msg("unable to make config dir to save backup")
-
-			return
-		}
-	}
-
-	for _, src := range configs {
-		dst := filepath.Join(baseDir, filepath.Base(src)+"."+ts)
-
-		sfi, err := os.Stat(src)
-		if err != nil {
-			log.Error().Err(err).Str("src", src).Msg("stat source file")
-
-			return
-		}
-
-		if !sfi.Mode().IsRegular() {
-			log.Error().Str("src", src).Msg("source is not a regular file")
-
-			return
-		}
-
-		in, err := os.Open(src)
-		if err != nil {
-			log.Error().Err(err).Str("src", src).Msg("opening source file")
-
-			return
-		}
-
-		out, err := os.Create(dst)
-		if err != nil {
-			log.Error().Err(err).Str("dst", dst).Msg("creating destination file")
-			in.Close()
-
-			return
-		}
-
-		if _, err := io.Copy(out, in); err != nil {
-			log.Error().Err(err).Msg("copying file contents")
-			in.Close()
-			out.Close()
-
-			return
-		}
-
-		in.Close()
-		out.Close()
-		log.Info().Str("src", src).Str("dst", dst).Msg("backed up original config file")
-	}
 }
